refactor pl usecase
