+++ src/main/kotlin/com/example/accounting/usecase/profit_and_loss/ListProfitAndLossUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.accounting.usecase.journal
+package com.example.accounting.usecase.profit_and_loss
 
 import com.example.accounting.domain.account.AccountRepository
 import com.example.accounting.domain.account.AccountType
@@ -6,9 +6,7 @@ import com.example.accounting.domain.common.DateRange
 import com.example.accounting.domain.journal.JournalRepository
 import com.example.accounting.domain.profit_and_loss.Pl
 import com.example.accounting.domain.profit_and_loss.ProfitAndLoss
-import org.springframework.cglib.core.Local
 import org.springframework.stereotype.Service
-import java.time.LocalDate
 import java.time.YearMonth
 
 @Service
@@ -19,18 +17,18 @@ class ListProfitAndLossUseCase(
     fun execute(
         month: YearMonth,
     ): Pl {
-        val range = DateRange.of(
+        val range = DateRange.Companion.of(
             month.atDay(1),
             month.atEndOfMonth(),
         )
 
         /* P/L に載せる科目(収益・費用)を取得 */
-        val plTypes = AccountType.getListForPL()
+        val plTypes = AccountType.Companion.getListForPL()
         val accounts = accountRepository.filterByAccountTypes(plTypes)
         /* 今月の仕訳を取得 */
         val journals = journalRepository.filterByDateRange(range)
 
-        val result =  ProfitAndLoss.create(
+        val result =  ProfitAndLoss.Companion.create(
             accounts,
             journals
         )
